Moved getAccumulatedObjects to LeftTuple
